Move api-mocks to test utils and remove unneccessary types file.
