+++ coral/src/app/features/topics/acl-request/TopicAclRequest.test.tsx
@@ -18,7 +18,7 @@ import {
   mockedResponseTopicTeamLiteral,
 } from "src/domain/topic/topic-api.msw";
 import api from "src/services/api";
-import { server } from "src/services/api-mocks/server";
+import { server } from "src/services/test-utils/api-mocks/server";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
 
 const mockedNavigate = jest.fn();
