+++ coral/src/app/features/topics/request/TopicRequest.test.tsx
@@ -10,7 +10,7 @@ import {
   mockgetTopicAdvancedConfigOptions,
 } from "src/domain/topic/topic-api.msw";
 import api from "src/services/api";
-import { server } from "src/services/api-mocks/server";
+import { server } from "src/services/test-utils/api-mocks/server";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
 import { objectHasProperty } from "src/services/type-utils";
