--- coral/src/services/api-mocks/types.ts
@@ -1,5 +1,0 @@
-import { SetupServerApi } from "msw/node";
-
-type MswInstance = SetupServerApi;
-
-export type { MswInstance };
